Create only one NumberingInstance per level in unordered lists
Track the maximum indentation level for unordered lists (<ul>) and create only one NumberingInstance per level. More is not nessecary, because they would not look different on the same level.
Too many  NumberingInstances are a problem because MS Word cannot cope with hundreds of them and will mess up all list in the rest of the document.